Tighten up usage output
Don't factor things in the maxWidth if they're gonna overrun anyway.

Only skip lines between options if either of them are likely to wrap to
multiple lines.

This means that a bunch of short option descriptions won't take up an
enormous amount of screen real-estate, to the point where the added
white space reduces legibility.